requirements.txt
# SIPaKMeD - Cervical Cancer Classification
# Core ML libraries
tensorflow>=2.10.0
tensorflow-hub>=0.12.0
keras>=2.10.0

# Data manipulation
numpy>=1.21.0
pandas>=1.4.0

# Image processing
opencv-python>=4.6.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Machine Learning utilities
scikit-learn>=1.1.0
imbalanced-learn>=0.9.0

# Progress bars
tqdm>=4.64.0

# System utilities
pathlib>=1.0.1